usability improvements; see details...
remove requirement to specify the maximum df expansion distance. this
will default to the radius of the largest sphere on the robot. allow this to be
overridden by the user by the old config element.

add ros param for specifying the mapping from joint group to collision
group to allow config to be reused between joint groups.